Move numbering logic from core to tool module
Refactored numbering-related classes and functions from src/bonsai/bonsai/core/numbering.py to src/bonsai/bonsai/tool/numbering.py. Updated imports in operator, prop, and ui modules to use bonsai.tool.numbering. Adjusted workspace keymap properties and extended the Numbering interface in core/tool.py. This improves modularity and separation of concerns between core and tool logic.
